A11y: Change Logical Order of Utility Nav
Utility nav was placed logically based on desktop location, but this lead to
issues for mobile, where the utility nav was technically before the Menu Toggle button.

Now logical order follows mobile display order, which still provides a nice sequence on desktop.

A major benefit on desktop is that the search field is not directly proceeded by the Search Toggle button, which it was not before.